Prefetch all pages
